+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.2.0 (unpublished)
+## 0.2.0
 
 * 💥 `onBeforeChange` is no longer called with the `nextElement` parameter.
 
@@ -6,6 +6,7 @@
 * If at least intro.js 2.8.0 is used, `false` can be returned in `onBeforeChange` to prevent transitioning to the next / previous step.
 * Add the `onPreventChange` callback called when a transition is prevented by returning `false` in `onBeforeChange`.
 * Add the `onBeforeExit` callback to prevent exiting the intro.
+* Update intro.js `import` statements to avoid a deprecation warning.
 
 ## 0.1.5